# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_financeos_api_internal_domain_entity.Category:
    properties:
      color:
        type: string
      created_at:
        type: string
      icon:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_system:
        type: boolean
      name:
        type: string
      parent_id:
        type: string
      type:
        description: income, expense, transfer
        type: string
      user_id:
        description: null = system category
        type: string
    type: object
...
  github_com_financeos_api_internal_domain_entity.FamilyGroup:
    properties:
      created_at:
        type: string
      id:
        type: string
      invite_code:
        type: string
      name:
        type: string
      owner_id:
        type: string
    type: object
  github_com_financeos_api_internal_domain_entity.FamilyMember:
    properties:
      group_id:
        type: string
      id:
        type: string
      joined_at:
        type: string
      permissions:
        additionalProperties: true
        type: object
      user_email:
        type: string
      user_id:
        type: string
      user_name:
        description: Joined fields
        type: string
    type: object
...
  github_com_financeos_api_internal_domain_entity.Transaction:
    properties:
      account_id:
        type: string
      account_name:
        description: Joined fields
        type: string
      ai_categorized:
        type: boolean
      ai_confidence:
        type: number
      amount:
        type: number
      category_color:
        type: string
      category_icon:
        type: string
      category_id:
        type: string
      category_name:
        type: string
      created_at:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      import_id:
        type: string
      notes:
        type: string
      recurrence_id:
        type: string
      tags:
        items:
          type: string
        type: array
      transfer_pair_id:
        type: string
      type:
        description: income, expense, transfer
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_financeos_api_internal_usecase.CreateCategoryRequest:
    properties:
      color:
        type: string
      icon:
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      parent_id:
        type: string
      type:
        enum:
        - income
        - expense
        - transfer
        type: string
    required:
    - name
    - type
    type: object
...
  github_com_financeos_api_internal_usecase.UpdateCategoryRequest:
    properties:
      color:
        type: string
      icon:
        type: string
      name:
        type: string
    type: object
...
  internal_handler.CashflowResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_handler.MonthlyCashflowItem'
        type: array
    type: object
  internal_handler.CategoryExpense:
    properties:
      amount:
        example: 350
        type: number
      category_id:
        example: uuid
        type: string
      category_name:
        example: Alimentação
        type: string
    type: object
  internal_handler.CategoryListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_financeos_api_internal_domain_entity.Category'
        type: array
    type: object
  internal_handler.CategoryResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_financeos_api_internal_domain_entity.Category'
    type: object
...
  internal_handler.DashboardOverviewData:
    properties:
      month_expense:
        example: 1500
        type: number
      month_income:
        example: 8000
        type: number
      net_balance:
        example: 22850
        type: number
      recent_transactions:
        items:
          $ref: '#/definitions/github_com_financeos_api_internal_domain_entity.Transaction'
        type: array
      top_categories:
        items:
          $ref: '#/definitions/internal_handler.CategoryExpense'
        type: array
    type: object
  internal_handler.DashboardOverviewResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handler.DashboardOverviewData'
    type: object
  internal_handler.ErrorDetail:
    properties:
      code:
        example: INVALID_INPUT
        type: string
      message:
        example: name is required
        type: string
    type: object
  internal_handler.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/internal_handler.ErrorDetail'
    type: object
  internal_handler.FamilyDashboardData:
    properties:
      members:
        items:
          $ref: '#/definitions/github_com_financeos_api_internal_domain_entity.FamilyMember'
        type: array
      total_expense:
        example: 3000
        type: number
      total_income:
        example: 16000
        type: number
    type: object
  internal_handler.FamilyDashboardResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handler.FamilyDashboardData'
    type: object
  internal_handler.FamilyGroupResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_financeos_api_internal_domain_entity.FamilyGroup'
    type: object
...
  internal_handler.InviteCodeData:
    properties:
      invite_code:
        example: ABCD1234
        type: string
    type: object
  internal_handler.InviteCodeResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handler.InviteCodeData'
    type: object
  internal_handler.MessageData:
    properties:
      message:
        example: logged out successfully
        type: string
    type: object
  internal_handler.MessageResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handler.MessageData'
    type: object
  internal_handler.MonthlyCashflowItem:
    properties:
      balance:
        example: 6500
        type: number
      expense:
        example: 1500
        type: number
      income:
        example: 8000
        type: number
      month:
        example: 4
        type: integer
      year:
        example: 2026
        type: integer
    type: object
...
host: localhost:8000
info:
  contact:
    name: FinanceOS Support
    url: https://github.com/financeos
  description: API completa para gestão financeira pessoal — contas, transações, orçamentos,
    investimentos, metas e IA.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: FinanceOS API
  version: 1.0.0
paths:
...
  /categories:
    get:
      description: Retorna categorias do sistema e categorias personalizadas do usuário
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.CategoryListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar categorias
      tags:
      - Categories
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados da categoria
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_financeos_api_internal_usecase.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar categoria
      tags:
      - Categories
  /categories/{id}:
    delete:
      parameters:
      - description: UUID da categoria
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remover categoria
      tags:
      - Categories
    put:
      consumes:
      - application/json
      parameters:
      - description: UUID da categoria
        in: path
        name: id
        required: true
        type: string
      - description: Campos a atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_financeos_api_internal_usecase.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.CategoryResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar categoria
      tags:
      - Categories
...
  /dashboard/cashflow:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.CashflowResponse'
      security:
      - BearerAuth: []
      summary: Cashflow dos últimos 12 meses
      tags:
      - Dashboard
  /dashboard/overview:
    get:
      description: Retorna saldo líquido, receitas, despesas e top categorias
      parameters:
      - description: 'Mês (1-12, default: atual)'
        in: query
        name: month
        type: integer
      - description: 'Ano (default: atual)'
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.DashboardOverviewResponse'
      security:
      - BearerAuth: []
      summary: Visão geral do mês
      tags:
      - Dashboard
  /family:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.FamilyGroupResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obter grupo familiar
      tags:
      - Family
    post:
      consumes:
      - application/json
      parameters:
      - description: Nome do grupo
        in: body
        name: body
        required: true
        schema:
          properties:
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.FamilyGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar grupo familiar
      tags:
      - Family
  /family/dashboard:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.FamilyDashboardResponse'
      security:
      - BearerAuth: []
      summary: Dashboard do grupo familiar
      tags:
      - Family
  /family/invite:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.InviteCodeResponse'
      security:
      - BearerAuth: []
      summary: Gerar código de convite
      tags:
      - Family
  /family/join:
    post:
      consumes:
      - application/json
      parameters:
      - description: Código de convite
        in: body
        name: body
        required: true
        schema:
          properties:
            invite_code:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.FamilyGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Entrar em grupo familiar via convite
      tags:
      - Family
  /family/members/{id}:
    delete:
      parameters:
      - description: UUID do membro
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Remover membro do grupo
      tags:
      - Family
...
securityDefinitions:
  BearerAuth:
    description: 'Digite: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"